refactor(api): merge manager webhooks
